Amplfiers virker måske?? Ikke testet

+++ PropagationPuzzle/Assets/Scripts/RaycastSound/SoundRay.cs
@@ -4,16 +4,16 @@ public struct SoundRay // Values
 {
     public Vector3 direction;
     public int reflections;
-    public int occlusions;
+    public float occlusions;
 
-    public SoundRay(Vector3 d, int r, int o)
+    public SoundRay(Vector3 d, int r, float o)
     {
         direction = d;
         reflections = r;
         occlusions = o;
 
     }
-    public void SetValues (Vector3 d, int r, int o)
+    public void SetValues (Vector3 d, int r, float o)
     {
         direction = d;
         reflections = r;
